AWS Improved Health Check fix.
When AWS Improved Health Check is enabled, it is possible for the status of the environment to never enter the "Grey" health state.

In particular, this can happen every time if the environment was previously in a Severe state, but it can sometimes happen when updating from a prior green state.