options to define vpc and subnets
vpc - id or name of the vpc. if not found failure. if found add to
option_settings
vpc_subnets - comma seperated list of ids or names from vpc subnets.
looks for subnets in previous defined vpc. if no vpc defined it looks
over all vpc. if not found failure. if found add to option_settings